requirements.txt
python-telegram-bot==12.4.1
beautifulsoup4==4.8.2
requests==2.22.0
lxml==4.5
# python_dateutil==2.8.1
